Dav Mount support in Terminal

+++ public/apps/terminal.tapp/scripts/rmdir.js
@@ -1,16 +1,20 @@
-function rmdir(args) {
+async function rmdir(args) {
 	if (args._raw.length <= 0) {
 		displayError("rmdir: missing operand");
 		createNewCommandInput();
 		return;
 	}
-	tb.sh.rm(`${path}/${args._raw}`, err => {
-		if (err) {
-			displayError(`rmdir: ${err.message}`);
-			createNewCommandInput();
-		} else {
-			createNewCommandInput();
-		}
-	});
+	if (path.includes("/mnt/")) {
+		displayError("TNSM rmdir: Removing directories from mounted drives is not supported by the webdav library at this time.");
+	} else {
+		tb.sh.rm(`${path}/${args._raw}`, err => {
+			if (err) {
+				displayError(`rmdir: ${err.message}`);
+				createNewCommandInput();
+			} else {
+				createNewCommandInput();
+			}
+		});
+	}
 }
 rmdir(args);
